Refactor how to stat app
