+++ chat-server/index.ts
@@ -1,6 +1,5 @@
-import App from './server';
-import http from 'http';
-import express from 'express';
+import App from './express-server';
+import {createServer, Server} from 'http';
 import './config';
 
 // Controllers
@@ -8,22 +7,8 @@ import { AuthController } from 'chat-server/src/auth';
 import { RoomController } from 'chat-server/src/room';
 import { ChatSocketController } from 'chat-server/src/socket-chat';
 
-async function startServer() {
-  const app = express();
-  const server = http.createServer(app);
-  const port = process.env.PORT;
 
-  new App(app, [new AuthController(), new RoomController()]);
-  new ChatSocketController(server);
-
-  await require('./loaders').default();
-  server.listen(port, () => {
-    console.log(`
-    #########################################
-    Server listening on port: ${port}
-    #########################################
-  `);
-  });
-}
-
-startServer();
+const app = new App([new AuthController(), new RoomController()]);
+const server = createServer(app.getServer());
+new ChatSocketController(server);
+app.listen(server);
