+++ chat-server/package.json
@@ -15,6 +15,7 @@
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "cookie-parser": "^1.4.4",
+    "cors": "^2.8.5",
     "envalid": "^5.0.0",
     "express": "^4.16.4",
     "express-validator": "^6.1.1",
@@ -30,6 +31,7 @@
     "@types/bcryptjs": "^2.4.2",
     "@types/body-parser": "^1.17.0",
     "@types/cookie-parser": "^1.4.1",
+    "@types/cors": "^2.8.5",
     "@types/express": "^4.16.1",
     "@types/jsonwebtoken": "^8.3.2",
     "@types/lodash": "^4.14.133",
